Moved app-specific classes to org.cirdles.topsoil.app
