+++ src/main/java/org/cirdles/topsoil/app/TopsoilMainWindow.java
@@ -51,7 +51,7 @@ import org.cirdles.topsoil.app.utils.TSVDatasetReader;
 import org.cirdles.topsoil.app.utils.TSVDatasetWriter;
 import org.cirdles.topsoil.app.utils.DatasetReader;
 import org.cirdles.topsoil.app.utils.DatasetWriter;
-import org.cirdles.topsoil.chart.ChartInitializationDialog;
+import org.cirdles.topsoil.app.chart.ChartInitializationDialog;
 import org.cirdles.topsoil.chart.JavaScriptChart;
 import org.cirdles.topsoil.data.DatasetManager;
 import org.cirdles.topsoil.data.TSVDatasetManager;
