+++ src/main/java/org/cirdles/topsoil/app/chart/ChartInitializationDialog.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.chart;
+package org.cirdles.topsoil.app.chart;
 
+import org.cirdles.topsoil.chart.JavaScriptChart;
 import org.cirdles.topsoil.data.Dataset;
 import org.controlsfx.dialog.Dialog;
