+++ src/main/java/org/cirdles/topsoil/app/chart/ChartInitializationView.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.chart;
+package org.cirdles.topsoil.app.chart;
 
 import org.cirdles.topsoil.app.*;
 import java.util.ArrayList;
