+++ src/test/java/org/cirdles/topsoil/app/UncertaintyVariableFormatTest.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.chart;
+package org.cirdles.topsoil.app;
 
+import org.cirdles.topsoil.app.UncertaintyVariableFormat;
 import static org.cirdles.topsoil.app.ExpressionType.*;
 import org.junit.Test;
 import static org.junit.Assert.*;
